Enemy VFX. Enemy canvas linked. Canvas lookat cam. Working on cam tween

+++ Minor Skilled/Assets/Scripts/Tank components/TankData.cs
@@ -48,7 +48,7 @@ public class TankData
         int currentArmor = 0;
         foreach (var part in TankParts)
         {
-            currentArmor += part.CurrentArmor;
+            currentArmor += (int)part.CurrentArmor;
         }
 
         return currentArmor;
